refactor: Move draft actions to ActionsViewModel and use refreshAsync
